Updated Frontend for Landing Page, Admin Login, Staff login, and Retractable Sidebar
Updated frontend of the three while also creating a new directory at public/img for the dswd logo and soc pen logo, feel free to change the src code of the three to your existing directory for the image,,

Edited the sidebar for the admin side to be retractable showing only the icons when retracted and hovering only the icons when retracted. Pressing the icons are also interactable redirecting you to the other tabs.